requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
httpx==0.26.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.0
stripe>=8.0.0
